Fix calculation of whitespace length in string literals.
To do this, introduces a new token type LiteralWhitespaceToken which is now used instead of WhitespaceToken inside string literals. It acts like a WhitespaceToken except that the token content is always output literally, rather than being replaced by a single space or a line break. This makes it easier to decide in the later stages whether the content length should be used or not to decide the total token length.