Support KDoc embedded in a modifier list, moving it to the beginning of the list.
Previously, KDoc between an annotation and a modifier keyword would not be recognized by the node scanner, since it was expected that KDoc be outside the modifier list. This adds support for KDoc to the modifier list directly so that it is moved to be prior to the list even when there is a modifier keyword on the declaration.

Fixes #4