Force a break after an annotation unless it's a simple annotation.
This is required by the [Kotlin coding conventions](https://kotlinlang.org/docs/coding-conventions.html#annotations). Previously, an annotation with a parameter would be left on the same line as what it was annotating if that would not violate the column limit. Now, a line break is always added after the annotation.

Fixes #17