Prefer to break within a parameter list if a function declaration does not fit on a line.
Previously, the formatter would break just before the return type instead if everything preceding the return type would fit on a line. This is technically not compliant with the [coding conventions](https://kotlinlang.org/docs/coding-conventions.html#functions).

Now, if the full function declaration does not fit on the line, the formatter will first break the parameter list if there are any parameters, and then break before the return type only if the parameter list is empty.

Fixes #18