Ignore intervening BeginToken's when calculating the length of a WhitespaceToken.
This ensures that the line breaker will not fail to introduce a line break and thereby break the column limit just because it couldn't find any further opportunities to insert a line break.

Also reworks the dot qualified expression scanner to ensure, in combination with the change above, that dot qualified expressions with really long reference names do not violate the column limit.

Fixes #20