Improve recognition of comments around if-else expressions.
Previously, comments at various points surrounding if-else would not be recognised by the auto-formatter, resulting in an error during formatting. This change ensures that such comments are recognised and formatted correctly.

Constructions of the form

    while (condition) // Comment
        {
            ...
        }

will be formatted by indenting the block as shown above, which one may see as ugly. This is best solved by putting the comment before the while statement:

    // Comment
    while (condition) {
        ...
    }

or after the opening brace:

    while (condition) { // Comment
        ...
    }

Handling the former case specially would be quite complex, so it's not specially supported. The same is true of else clauses in if-expressions.

Fixes #22